dns: write original DNS servers to /var/run/NetworkManager/stub-resolv.conf
When a DNS plugin is enabled (like "main.dns=dnsmasq" or "main.dns=systemd-resolved"),
the name servers announced to the rc-manager are coerced to be 127.0.0.1
or 127.0.0.53.

Depending on the "main.rc-manager" setting, also /etc/resolv.conf
contains only this coerced name server to the local caching service.

Likewise, the /var/run/NetworkManager/resolv.conf file, which is essentially
what we write to the main.rc-manager.

Write a new file /var/run/NetworkManager/stub-resolv.conf, which contains
the original name servers, uncoerced. This file is always written.

The effect is, when one enables "main.dns=systemd-resolved", then there
is still a file "stub-resolv.conf" with the content as with
"main.dns=default".

The stub-resolv.conf may be a possible solution, when one wants
NetworkManager to update systemd-resolved, but still have a regular
/etc/resolv.conf [1]. For that, the user would configure

    [main]
    dns=systemd-resolved
    rc-manager=unmanaged

and symlink /etc/resolv.conf to /var/run/NetworkManager/resolv.conf.
But it's not necessarily the only solution and does not precule options
for updating systemd-resolved in combination with other DNS plugins.

[1] https://gitlab.freedesktop.org/NetworkManager/NetworkManager/issues/20